Restructures dronekit.lib to be just dronekit.
