+++ docs/examples/follow_me.rst
@@ -108,7 +108,7 @@ Most of the example should be fairly familiar as it uses the same code as other 
 :ref:`taking off <taking-off>`, and closing the vehicle object. 
 
 The example-specific code is shown below.  All this does is attempt to get a gps socket and read the location in a two second loop. If it is successful it 
-reports the value and uses :py:func:`Vehicle.commands.goto <dronekit.lib.CommandSequence.goto>` to move to the new position. The loop exits when 
+reports the value and uses :py:func:`Vehicle.commands.goto <dronekit.CommandSequence.goto>` to move to the new position. The loop exits when 
 the mode is changed. 
 
 .. code-block:: python
