+++ docs/guide/auto_mode.rst
@@ -61,12 +61,12 @@ wiki topic provides a more detailed overview of commands.
 Download current mission
 ========================
 
-The mission commands for a vehicle are accessed using the :py:attr:`Vehicle.commands <dronekit.lib.Vehicle.commands>` 
-attribute. The attribute is of type :py:class:`CommandSequence <dronekit.lib.CommandSequence>`, a class that provides ‘array style’ indexed access to the 
+The mission commands for a vehicle are accessed using the :py:attr:`Vehicle.commands <dronekit.Vehicle.commands>` 
+attribute. The attribute is of type :py:class:`CommandSequence <dronekit.CommandSequence>`, a class that provides ‘array style’ indexed access to the 
 waypoints which make up the mission.
 
-Waypoints are not downloaded from vehicle until :py:func:`download() <dronekit.lib.CommandSequence.download>` is called. The download is asynchronous; 
-use :py:func:`wait_ready() <dronekit.lib.CommandSequence.wait_ready>` to block your thread until the download is complete:
+Waypoints are not downloaded from vehicle until :py:func:`download() <dronekit.CommandSequence.download>` is called. The download is asynchronous; 
+use :py:func:`wait_ready() <dronekit.CommandSequence.wait_ready>` to block your thread until the download is complete:
 
 .. code:: python
 
@@ -80,7 +80,7 @@ use :py:func:`wait_ready() <dronekit.lib.CommandSequence.wait_ready>` to block y
 
 .. note::
 
-    In DroneKit-Python 2 :py:attr:`Vehicle.commands <dronekit.lib.Vehicle.commands>` contains just the editable waypoints (in version 1.x, the 
+    In DroneKit-Python 2 :py:attr:`Vehicle.commands <dronekit.Vehicle.commands>` contains just the editable waypoints (in version 1.x, the 
     commands included the non-editable home location as the first item).
 
 
@@ -90,8 +90,8 @@ use :py:func:`wait_ready() <dronekit.lib.CommandSequence.wait_ready>` to block y
 Clearing current mission
 ========================
 
-To clear a mission you call :py:func:`clear() <dronekit.lib.CommandSequence.clear>` and then 
-:py:func:`Vehicle.commands.upload() <dronekit.lib.Vehicle.commands.upload>` (to upload the changes to the vehicle):
+To clear a mission you call :py:func:`clear() <dronekit.CommandSequence.clear>` and then 
+:py:func:`Vehicle.commands.upload() <dronekit.Vehicle.commands.upload>` (to upload the changes to the vehicle):
 
 .. code:: python
 
@@ -118,10 +118,10 @@ Creating/adding mission commands
 ================================
 
 After :ref:`downloading <auto_mode_download_mission>` or :ref:`clearing <auto_mode_clear_mission>` a mission new commands 
-can be added and uploaded to the vehicle. Commands are added to the mission using :py:func:`add() <dronekit.lib.CommandSequence.add>`
-and are sent to the vehicle (either individually or in batches) using :py:func:`upload() <dronekit.lib.Vehicle.commands.upload>`.
+can be added and uploaded to the vehicle. Commands are added to the mission using :py:func:`add() <dronekit.CommandSequence.add>`
+and are sent to the vehicle (either individually or in batches) using :py:func:`upload() <dronekit.Vehicle.commands.upload>`.
 
-Each command is packaged in a :py:class:`Command <dronekit.lib.Command>` object (see that class for the order/meaning of the parameters). 
+Each command is packaged in a :py:class:`Command <dronekit.Command>` object (see that class for the order/meaning of the parameters). 
 The supported commands for each vehicle are :ref:`linked above <auto_mode_supported_commands>`. 
 
 
@@ -187,7 +187,7 @@ modify them as needed, then clear ``Vehicle.commands`` and upload the list as a 
 
 
 The changes are not guaranteed to be complete until 
-:py:func:`upload() <dronekit.lib.Vehicle.commands.upload>` is called on the parent ``Vehicle.commands`` object.
+:py:func:`upload() <dronekit.Vehicle.commands.upload>` is called on the parent ``Vehicle.commands`` object.
 
 
 .. _auto_mode_monitoring_controlling: 
@@ -219,7 +219,7 @@ AUTO mode the mission will either restart at the beginning or resume at the curr
 `MIS_RESTART <http://copter.ardupilot.com/wiki/arducopter-parameters/#mission_restart_when_entering_auto_mode_mis_restart>`_ 
 parameter (available on all vehicle types).
 
-You can monitor the progress of the mission by polling the :py:func:`Vehicle.commands.next <dronekit.lib.CommandSequence.next>` attribute
+You can monitor the progress of the mission by polling the :py:func:`Vehicle.commands.next <dronekit.CommandSequence.next>` attribute
 to get the current command number. You can also change the current command by setting the attribute to the desired command number.
 
 .. code:: python
@@ -245,7 +245,7 @@ AUTO mode to start it running again.
 Currently there is no notification in DroneKit when a mission completes. If you need to detect mission end (in order
 to perform some other operation) then you can either:
 
-* Add a dummy mission command and poll :py:func:`Vehicle.commands.next <dronekit.lib.CommandSequence.next>` for the 
+* Add a dummy mission command and poll :py:func:`Vehicle.commands.next <dronekit.CommandSequence.next>` for the 
   transition to the final command, or
 * Compare the current position to the target position in the final waypoint.
 
@@ -292,10 +292,10 @@ Adding mission commands is discussed :ref:`here in the guide <auto_mode_adding_c
             vehicle.commands.upload()
 
 
-``readmission()`` reads a mission from the specified file and returns a list of :py:class:`Command <dronekit.lib.Command>` objects. 
+``readmission()`` reads a mission from the specified file and returns a list of :py:class:`Command <dronekit.Command>` objects. 
 
 Each line is split up. The first line is used to test whether the file has the correct (stated) format. 
-For subsequent lines the values are stored in a :py:class:`Command <dronekit.lib.Command>` object 
+For subsequent lines the values are stored in a :py:class:`Command <dronekit.Command>` object 
 (the values are first cast to the correct ``float`` and ``int`` types for their associated parameters).
 The commands are added to a list which is returned by the function.
   
@@ -357,7 +357,7 @@ It uses ``download_mission()`` (below) to get them mission, and then writes the 
         with open(aFileName, 'w') as file_:
             file_.write(output)      
 
-``download_mission()`` downloads the :py:attr:`Vehicle.commands <dronekit.lib.Vehicle.commands>` from the vehicle and 
+``download_mission()`` downloads the :py:attr:`Vehicle.commands <dronekit.Vehicle.commands>` from the vehicle and 
 adds them to a list. Downloading mission is discussed :ref:`in the guide <auto_mode_download_mission>`.
 
 .. code:: python
@@ -404,7 +404,7 @@ Get distance to waypoint
         distancetopoint = get_distance_metres(vehicle.location.global_frame, targetWaypointLocation)
         return distancetopoint
 
-The function determines the current target waypoint number with :py:func:`Vehicle.commands.next <dronekit.lib.CommandSequence.next>`
+The function determines the current target waypoint number with :py:func:`Vehicle.commands.next <dronekit.CommandSequence.next>`
 and uses it to index the commands to get the latitude, longitude and altitude of the target waypoint. The ``get_distance_metres()`` function
 (see :ref:`guided_mode_copter_useful_conversion_functions`) is then used to calculate and return the (horizontal) distance 
 from the current vehicle location.
