+++ dronekit/lib.py
@@ -1,0 +1,2 @@
+# Backwards compatibility
+from dronekit import *
