+++ dronekit/test/sitl/test_channels.py
@@ -1,5 +1,3 @@
-from dronekit import connect
-from dronekit.lib import VehicleMode
 from pymavlink import mavutil
 import time
 from dronekit import connect, VehicleMode, LocationGlobal
