+++ dronekit/test/unit/test_api.py
@@ -1,7 +1,6 @@
 import mock
 from mock import MagicMock
-import dronekit
-from dronekit.lib import VehicleMode
+from dronekit import VehicleMode
 from nose.tools import assert_equals, assert_not_equals
 
 def test_vehicle_mode_eq():
