+++ examples/gcs/microgcs.py
@@ -2,8 +2,7 @@
 # This is a small example of the python drone API - an ultra minimal GCS
 #
 
-from dronekit import connect
-from dronekit.lib import VehicleMode
+from dronekit import connect, VehicleMode
 from pymavlink import mavutil
 from Tkinter import *
